+++ src/server/routers/lambda/index.ts
@@ -52,6 +52,7 @@ import { notebookRouter } from './notebook';
 import { notificationRouter } from './notification';
 import { oauthDeviceFlowRouter } from './oauthDeviceFlow';
 import { platformAnalyticsRouter } from './platformAnalytics';
+import { platformAuditRouter } from './platformAudit';
 import { pluginRouter } from './plugin';
 import { ragEvalRouter } from './ragEval';
 import { recentRouter } from './recent';
@@ -116,6 +117,7 @@ export const lambdaRouter = router({
   notification: notificationRouter,
   oauthDeviceFlow: oauthDeviceFlowRouter,
   plugin: pluginRouter,
+  platformAudit: platformAuditRouter,
   platformAnalytics: platformAnalyticsRouter,
   ragEval: ragEvalRouter,
   recent: recentRouter,